refactor: re-arrange file/folder structure for expo router
