+++ app/_layout.tsx
@@ -1,0 +1,10 @@
+import { SafeAreaView } from "react-native-safe-area-context";
+import { Slot } from "expo-router";
+
+export default function AppLayout() {
+  return (
+    <SafeAreaView>
+      <Slot />
+    </SafeAreaView>
+  );
+}
