+++ app/index.tsx
@@ -1,0 +1,11 @@
+import { View, Text } from "react-native";
+
+const Home = () => {
+  return (
+    <View>
+      <Text>Hello world</Text>
+    </View>
+  );
+};
+
+export default Home;
