+++ package.json
@@ -26,7 +26,8 @@
     "react-native-screens": "~3.20.0",
     "expo-linking": "~4.0.1",
     "expo-constants": "~14.2.1",
-    "react-native-gesture-handler": "~2.9.0"
+    "react-native-gesture-handler": "~2.9.0",
+    "react-native-web": "~0.18.10"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
