+++ screens/Home.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { View, StyleSheet } from "react-native";
+import { View } from "react-native";
 import Input from "../components/Input";
 import List from "../components/List";
